+++ cypress/integration/global_spec.js
@@ -125,6 +125,7 @@ describe('해외증시', () => {
 
   describe('해외 대표 종목', () => {
     it('차트와 뉴스, 종목 리스트를 보여준다.', () => {
+      cy.stubImages();
       visit();
       // DOW와 NASDAQ 요청 대기
       cy.wait('@apiOverseasRepresentativeStock');
